Thirty-two commits from an external critique of the codebase. Every claim was
verified against the tree first: the headline one (a `last_activity_ns` data
race) was a misread — all accesses are owner-thread-pinned — but the sweep
found a real cross-thread bug the critique missed, and three of its structural
complaints turned out worse than stated.

Concurrency
  - quic/wake.zig `Handle.fds` was read and written by `request()` from any
    thread with no guard against `deinit()` nulling and closing them on the
    owner: the write-to-a-recycled-descriptor shape TCP had already fixed.
    Now mutex-guarded, with a cross-thread stress suite.
  - A ThreadSanitizer lane (Linux-only; libtsan SIGSEGVs on darwin at this
    pin) over the threaded transport suites, plus a nightly soak variant.
  - QUIC gained TCP's thread-check parity: a release opt-in and loop-thread
    assertions on the session accessors. QUIC `sendFrame` stays deliberately
    unchecked — its enqueue is mutex-guarded, so the asymmetry is by design.
  - `.closing` observer events now fire on the owner/loop thread only, and
    events.zig documents the observer threading contract that never existed.

Decomposition — no behavior change, frozen surface byte-identical throughout
  - peer/mod.zig 14,059 -> 5,807 lines (-59%) across the full P0-P12 ladder,
    into comptime-generic sibling modules one directory level deep. The
    deinit body moved as ONE unit; its teardown order is load-bearing.
  - build.zig 1,510 -> 14 lines behind build/, with `zig build -l` proven
    byte-identical and `build/` added to the package `.paths` whitelist.
  - The codegen splits were attempted and REVERTED: Zig's file-scoped privacy
    would force ~11 private helpers public, and `p("capnpc-zig.codegen")`
    sweeps generator.zig's whole pub surface into the FROZEN tier. Recorded
    at the rule; unblocking it is an API decision, not a refactor.

Evidence
  - Structural fuzz for the L3/L4 peer surface, the QUIC framers, and
    persistence restore — which found and fixed a real self-loopback stash
    leak on duplicate answer ids.
  - The experimental API snapshots are gated in CI for the first time. That
    required fixing the root cause: stored thread ids widened to u64, since
    `std.Thread.Id` is u32 on linux/windows and u64 on darwin and @typeName
    renders the resolved integer. This restores the staleness check dropped
    in 25cf197 as "not platform-stable" — it is stable now.
  - The cross-target matrix compiles on all four targets for the first time:
    64-bit atomics in the soak harness broke 32-bit x86 (pre-existing), and a
    u32-vs-u64 Thread.Id assertion broke the Windows test compile.

Supply chain
  - quic-zig is pinned to the immutable v0.10.1 tag rather than a raw SHA.

Verification: full local release-preflight green (exit 0) — 1376/1376 tests,
QUIC evidence 61/61 zero-skip in Debug and ReleaseSafe, e2e against Docker
cpp/go/python/rust, clean-room consumers for all three library roots,
bench-check 19/19 (RPC round-trip p50 24.1us vs a 50.3us baseline — the
decomposition cost nothing).

+++ tools/soak_rpc.zig
@@ -51,17 +51,23 @@ const Config = struct {
     mem_growth_pct: f64 = 25.0,
 };
 
+/// Counters are `usize`, not `u64`, so this harness cross-compiles for
+/// 32-bit targets: `@atomicLoad`/`@atomicRmw` reject operands wider than the
+/// pointer width, and `zig build check-compile -Dtarget=x86-linux-gnu` builds
+/// this file. `usize` is u64 on every machine the soak actually runs on, so
+/// the counters are unchanged there; the 32-bit build is compile-only rot
+/// detection, never an execution target.
 const Totals = struct {
-    sessions: std.atomic.Value(u64) = std.atomic.Value(u64).init(0),
-    calls_ok: std.atomic.Value(u64) = std.atomic.Value(u64).init(0),
-    calls_cancelled: std.atomic.Value(u64) = std.atomic.Value(u64).init(0),
-    chaos_closes: std.atomic.Value(u64) = std.atomic.Value(u64).init(0),
-    transport_errors: std.atomic.Value(u64) = std.atomic.Value(u64).init(0),
-    expected_disconnects: std.atomic.Value(u64) = std.atomic.Value(u64).init(0),
+    sessions: std.atomic.Value(usize) = std.atomic.Value(usize).init(0),
+    calls_ok: std.atomic.Value(usize) = std.atomic.Value(usize).init(0),
+    calls_cancelled: std.atomic.Value(usize) = std.atomic.Value(usize).init(0),
+    chaos_closes: std.atomic.Value(usize) = std.atomic.Value(usize).init(0),
+    transport_errors: std.atomic.Value(usize) = std.atomic.Value(usize).init(0),
+    expected_disconnects: std.atomic.Value(usize) = std.atomic.Value(usize).init(0),
     // Disconnect-reason Returns on non-chaos sessions: legitimate at high peer
     // counts (idle-timeout drops under contention), not a correctness failure.
-    contention_disconnects: std.atomic.Value(u64) = std.atomic.Value(u64).init(0),
-    unexpected_exceptions: std.atomic.Value(u64) = std.atomic.Value(u64).init(0),
+    contention_disconnects: std.atomic.Value(usize) = std.atomic.Value(usize).init(0),
+    unexpected_exceptions: std.atomic.Value(usize) = std.atomic.Value(usize).init(0),
 };
 
 fn nowNs(io: std.Io) i64 {
@@ -82,8 +88,8 @@ fn nowNsU(io: std.Io) u64 {
 
 const CountingAllocator = struct {
     backing: std.mem.Allocator,
-    allocated: std.atomic.Value(u64) = std.atomic.Value(u64).init(0),
-    freed: std.atomic.Value(u64) = std.atomic.Value(u64).init(0),
+    allocated: std.atomic.Value(usize) = std.atomic.Value(usize).init(0),
+    freed: std.atomic.Value(usize) = std.atomic.Value(usize).init(0),
 
     const vtable = std.mem.Allocator.VTable{
         .alloc = alloc,
@@ -102,7 +108,7 @@ const CountingAllocator = struct {
 
     /// Live bytes = allocated − freed. Saturating so a transient read that
     /// observes freed slightly ahead of allocated cannot underflow.
-    fn liveBytes(self: *const CountingAllocator) u64 {
+    fn liveBytes(self: *const CountingAllocator) usize {
         return self.allocated.load(.monotonic) -| self.freed.load(.monotonic);
     }
